+++ android/src/main/java/ru/innim/flutter_login_vk/FlutterError.java
@@ -1,7 +1,5 @@
 package ru.innim.flutter_login_vk;
 
-import com.vk.sdk.api.VKError;
-
 public class FlutterError {
     String code;
     String message;