+++ android/src/main/java/ru/innim/flutter_login_vk/Results.java
@@ -1,8 +1,8 @@
 package ru.innim.flutter_login_vk;
 
-import com.vk.sdk.VKAccessToken;
-import com.vk.sdk.api.VKError;
-import com.vk.sdk.api.model.VKApiUser;
+import com.vk.api.sdk.auth.VKAccessToken;
+import com.vk.sdk.api.base.dto.BaseBoolInt;
+import com.vk.sdk.api.users.dto.UsersUserXtrCounters;
 
 import java.util.HashMap;
 
@@ -26,41 +26,38 @@ public class Results {
     }
 
     public static HashMap<String, Object> error(final VKError error) {
-
         return new HashMap<String, Object>() {{
             put("apiCode", error.errorCode);
-            put("message", error.errorReason);
+            put("message", error.errorMsg);
         }};
     }
 
     public static HashMap<String, Object> accessToken(final VKAccessToken accessToken) {
         if (accessToken == null)
             return null;
-
         return new HashMap<String, Object>() {{
-            put("token", accessToken.accessToken);
-            put("userId", accessToken.userId);
-            put("created", accessToken.created);
-            put("expiresIn", accessToken.expiresIn);
-            put("email", accessToken.email);
-            put("httpsRequired", accessToken.httpsRequired);
-            put("secret", accessToken.secret);
+            put("token", accessToken.getAccessToken());
+            put("userId", String.valueOf(accessToken.getUserId()));
+            put("created", accessToken.getCreated());
+            put("email", accessToken.getEmail());
+            put("isValid", accessToken.isValid());
+            put("secret", accessToken.getSecret());
         }};
     }
 
-    public static HashMap<String, Object> userProfile(final VKApiUser user) {
+    public static HashMap<String, Object> userProfile(final UsersUserXtrCounters user) {
         if (user == null)
             return null;
 
         return new HashMap<String, Object>() {{
-            put("userId", user.getId());
-            put("firstName", user.first_name);
-            put("lastName", user.last_name);
-            put("online", user.online);
-            put("onlineMobile", user.online_mobile);
-            put("photo50", user.photo_50);
-            put("photo100", user.photo_100);
-            put("photo200", user.photo_200);
+            put("userId", user.getId().getValue());
+            put("firstName", user.getFirstName());
+            put("lastName", user.getLastName());
+            put("online", user.getOnline() == BaseBoolInt.YES);
+            put("onlineMobile", user.getOnlineMobile() == BaseBoolInt.YES);
+            put("photo50", user.getPhoto50());
+            put("photo100", user.getPhoto100());
+            put("photo200", user.getPhoto200());
         }};
     }
 }